refactor(proof-harnesses): extract shared utilities to proof-common.js to fix SonarCloud quality gate
Reduce 33.4% duplication between prove-vlt.js and prove-hf-context.js by
extracting check, ensureDir, patternMatches, runProofSuites, and
printReportAndExit into scripts/proof-common.js. Both harnesses now import
from proof-common.js, keeping their domain-specific test suites intact.

Fixes SonarCloud quality gate failures (duplication ≤ 3%, coverage ≥ 80%,
reliability A rating) on PR #3218.
